+++ components/shared/darkModeSwitcher.tsx
@@ -1,6 +1,5 @@
 "use client"
 
-import * as React from "react"
 import { MoonIcon, SunIcon } from "@radix-ui/react-icons"
 import { useTheme } from "next-themes"
 
@@ -15,15 +14,15 @@ import useExistingTheme from "@/hooks/useDarkMode"
 
 export function ModeToggle() {
     const { setTheme } = useTheme();
-    const ExistingTheme =  useExistingTheme();
-    if(ExistingTheme){
+    const ExistingTheme = useExistingTheme();
+    if (ExistingTheme) {
         setTheme(ExistingTheme);
     }
-    
+
     return (
         <DropdownMenu>
             <DropdownMenuTrigger asChild>
-                <Button variant="outline" size="icon">
+                <Button variant="secondary" size="icon">
                     <SunIcon className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
                     <MoonIcon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
                     <span className="sr-only">Toggle theme</span>